2010-02-07 02:02 UTC+0100 Viktor Szakats (harbour.01 syenar.hu)   * src/pp/ppcore.c   * src/vm/macro.c   * src/vm/runner.c   * src/vm/codebloc.c   * src/vm/garbage.c   * src/vm/hvm.c   * src/vm/maindllp.c   * src/vm/fm.c   * src/vm/extrap.c   * src/vm/memvars.c   * src/common/expropt1.c   * src/common/expropt2.c   * src/common/hbarch.c   * src/common/hbmem.c   * src/nortl/nortl.c   * src/rtl/gtdos/gtdos.c   * src/rtl/gtwin/gtwin.c   * src/rtl/gtxwc/gtxwc.h   * src/rtl/gtxwc/gtxwc.c   * src/rtl/gtcrs/gtcrs.c   * src/rtl/gtstd/gtstd.c   * src/rtl/gttrm/gttrm.c   * src/rtl/gtcgi/gtcgi.c   * src/rtl/gtapi.c   * src/rtl/gtos2/gtos2.c   * src/rtl/hbgtcore.c   * src/rtl/gtsln/gtsln.c   * src/rtl/gtsln/kbsln.c   * src/rtl/gtsln/gtsln.h   * src/rtl/gtpca/gtpca.c   * src/rtl/gt_tpl/gt_tpl.c   * src/rtl/gtwvt/gtwvt.h   * src/rtl/gtwvt/gtwvt.c   * src/rdd/dbfntx/dbfntx1.c   * src/rdd/dbsql.c   * src/rdd/hbdbsort.c   * src/rdd/workarea.c   * src/rdd/sdf1.c   * src/rdd/dbfnsx/dbfnsx1.c   * src/rdd/dbfcdx/dbfcdx1.c   * src/rdd/delim1.c   * src/rdd/dbf1.c   * src/rdd/hsx/hsx.c   * src/rdd/usrrdd/usrrdd.c   * src/rdd/dbffpt/dbffpt1.c   * src/rdd/hbsix/sxsem.c   * src/rdd/hbsix/sxcompr.c   * src/compiler/ppcomp.c   * src/compiler/hbmain.c   * src/compiler/hbfix.c   * src/compiler/hbdbginf.c   * src/compiler/hbpcode.c   * src/compiler/genhrb.c   * src/compiler/hbdead.c   * src/compiler/genobj32.c   * src/compiler/genc.c   * src/compiler/hbopt.c   * src/compiler/gencc.c   * src/compiler/hbcmplib.c   * src/compiler/hblbl.c   * src/compiler/harbour.yyc   * src/compiler/harbour.y   * src/compiler/harbour.yyh   * tests/bldtest/bldtest.c   * tests/tstgtapi.c   * include/hbpp.h   * include/hbrdddbf.h   * include/hbvmpub.h   * include/hbrddnsx.h   * include/hbgtcore.h   * include/hbxvm.h   * include/hbdefs.h   * include/hbmacro.h   * include/hbtypes.h   * include/hbrddsdf.h   * include/hbapi.h   * include/hbrddntx.h   * include/hbvm.h   * include/hbcomp.h   * include/hbapirdd.h   * include/hbcompdf.h   * include/hbrdddel.h   * include/hbexpra.c   * include/hbexprop.h   * include/hbdbf.h   * include/hbdbsort.h   * include/hbapigt.h   * include/hbrddcdx.h   * include/hbrddfpt.h   * include/hbexprb.c   * contrib/gtalleg/gtalleg.c   * contrib/hbct/ctwin.c   * contrib/hbct/ascpos.c   * contrib/hbct/screen1.c   * contrib/hbct/ctcrypt.c   * contrib/xhb/hbxml.c   * contrib/xhb/xhbis.c   * contrib/xhb/fparse.c   * contrib/xhb/xhbat.c   * contrib/xhb/xhbcopyf.c   * contrib/xhb/cstructc.c   * contrib/hbqt/gtqtc/gtqtc.cpp   * contrib/hbqt/gtqtc/gtqtc.h   * contrib/xpp/philesx.c   * contrib/hbnf/ftshadow.c   * contrib/hbnf/ftattr.c   * contrib/hbnetio/netiocli.c   * contrib/hbnetio/netiosrv.c   * contrib/rddsql/sddmy/mysqldd.c   * contrib/rddsql/sddpg/pgsqldd.c   * contrib/rddsql/sddfb/fbirddd.c   * contrib/rddsql/sddodbc/odbcdd.c   * contrib/rddsql/hbrddsql.h   * contrib/rddsql/sqlbase.c   * contrib/rddsql/sqlmix.c   * contrib/gtwvg/gtwvg.c   * contrib/gtwvg/wvggui.c   * contrib/gtwvg/wvgcore.c   * contrib/gtwvg/wvgwing.c   * contrib/rddads/adsx.c   * contrib/rddads/rddads.h   * contrib/rddads/ads1.c   * contrib/hbtip/utils.c   * contrib/hbwin/win_regc.c   * contrib/hbwin/win_prn2.c   * contrib/hbbmcdx/bmdbfcdx.c   * contrib/hbbmcdx/hbbmcdx.h   * contrib/hbbtree/hb_btree.c     * BYTE -> HB_BYTE except where BYTE was used in OS-specific context.     * Some formatting.     ; I didn't realign structure definitions in hbrdd*.h yet.
  * contrib/hbwin/win_regc.c
    * Minor type cleanup.

  * src/rtl/gt_tpl/gt_tpl.c
    * BYTE -> int for color.

  * examples/hbapollo/apollo.c
    % Deleted MAX_STR_LEN.

+++ harbour/contrib/hbnf/ftshadow.c
@@ -67,5 +67,5 @@ HB_FUNC( FT_SETATTR )
                       hb_parni( 2 ),
                       hb_parni( 3 ),
                       hb_parni( 4 ),
-                      ( BYTE ) hb_parni( 5 ) );
+                      hb_parni( 5 ) );
 }
